Updates to Strange2 and Modez
-Commented some functions
-Updated Bank 6 Modez
-Added a reset for the Bank 5 Attracktorz

+++ LeftBrain/PedalBrainz_LeftBrain_Software/v2/PedalBrainz_LeftBrain_Software_v2/PedalBrainz_LeftBrain_Software_v2.ino
@@ -95,7 +95,7 @@ int greenValue = 107;
 int fadeValueScaled;
 int scaledBright;
 
-int fadeSpeed = 13;
+float fadeSpeed = 13;
 int fullDelay = 0;
 int darkDelay = 0;
